Rename quest test package
