+++ android/quest/src/test/java/org/smartregister/fhircore/quest/QuestApplicationTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.smartregister.fhirecore.quest
+package org.smartregister.fhircore.quest
 
 import androidx.test.core.app.ApplicationProvider
 import org.hl7.fhir.r4.model.ResourceType
@@ -22,9 +22,8 @@ import org.junit.Assert
 import org.junit.Test
 import org.robolectric.annotation.Config
 import org.smartregister.fhircore.engine.configuration.app.ConfigurableApplication
-import org.smartregister.fhircore.quest.QuestApplication
-import org.smartregister.fhirecore.quest.robolectric.RobolectricTest
-import org.smartregister.fhirecore.quest.shadow.QuestApplicationShadow
+import org.smartregister.fhircore.quest.robolectric.RobolectricTest
+import org.smartregister.fhircore.quest.shadow.QuestApplicationShadow
 
 @Config(shadows = [QuestApplicationShadow::class])
 class QuestApplicationTest : RobolectricTest() {
