+++ android/quest/src/test/java/org/smartregister/fhircore/quest/robolectric/RobolectricTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.smartregister.fhirecore.quest.robolectric
+package org.smartregister.fhircore.quest.robolectric
 
 import android.os.Build
 import androidx.lifecycle.LiveData
