+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ui/patient/details/QuestPatientTestResultActivityTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.smartregister.fhirecore.quest.ui.patient.details
+package org.smartregister.fhircore.quest.ui.patient.details
 
 import android.app.Activity
 import io.mockk.every
@@ -25,9 +25,8 @@ import org.junit.Test
 import org.robolectric.Robolectric
 import org.robolectric.annotation.Config
 import org.robolectric.util.ReflectionHelpers
-import org.smartregister.fhircore.quest.ui.patient.details.QuestPatientTestResultActivity
-import org.smartregister.fhirecore.quest.robolectric.ActivityRobolectricTest
-import org.smartregister.fhirecore.quest.shadow.QuestApplicationShadow
+import org.smartregister.fhircore.quest.robolectric.ActivityRobolectricTest
+import org.smartregister.fhircore.quest.shadow.QuestApplicationShadow
 
 @Config(shadows = [QuestApplicationShadow::class])
 class QuestPatientTestResultActivityTest : ActivityRobolectricTest() {
